Delete kani::expect_fail and associated hook
