+++ tests/expected/arbitrary/enums/expected
@@ -1,0 +1,48 @@
+Checking harness check_random...
+
+Status: SATISFIED\
+Description: "This should be reachable"\
+main.rs:67:1 in function check_random
+
+Status: SATISFIED\
+Description: "This should be reachable"\
+main.rs:67:1 in function check_random
+
+Status: SATISFIED\
+Description: "This should be reachable"\
+main.rs:67:1 in function check_random
+
+Checking harness check_basic...
+
+Status: SATISFIED\
+Description: "This should be reachable"\
+main.rs:65:1 in function check_basic
+
+Status: SATISFIED\
+Description: "This should be reachable"\
+main.rs:65:1 in function check_basic
+
+Status: SATISFIED\
+Description: "This should be reachable"\
+main.rs:65:1 in function check_basic
+
+Checking harness check_continuous...
+
+Status: SATISFIED\
+Description: "This should be reachable"\
+main.rs:66:1 in function check_continuous
+
+Status: SATISFIED\
+Description: "This should be reachable"\
+main.rs:66:1 in function check_continuous
+
+Status: SATISFIED\
+Description: "This should be reachable"\
+main.rs:66:1 in function check_continuous
+
+ ** 3 of 3 cover properties satisfied
+
+
+VERIFICATION:- SUCCESSFUL
+
+Complete - 3 successfully verified harnesses, 0 failures, 3 total.
