+++ tests/expected/arbitrary/integers/expected
@@ -1,0 +1,222 @@
+Checking harness check_u128...
+
+Status: SATISFIED\
+Description: "This may be true"\
+main.rs:31:1 in function check_u128
+
+Status: SATISFIED\
+Description: "This may also be true"\
+main.rs:31:1 in function check_u128
+
+Status: SATISFIED\
+Description: "v1 may be MAX"\
+main.rs:31:1 in function check_u128
+
+Status: SATISFIED\
+Description: "v1 may be MIN"\
+main.rs:31:1 in function check_u128
+
+Checking harness check_u32...
+
+Status: SATISFIED\
+Description: "This may be true"\
+main.rs:29:1 in function check_u32
+
+Status: SATISFIED\
+Description: "This may also be true"\
+main.rs:29:1 in function check_u32
+
+Status: SATISFIED\
+Description: "v1 may be MAX"\
+main.rs:29:1 in function check_u32
+
+Status: SATISFIED\
+Description: "v1 may be MIN"\
+main.rs:29:1 in function check_u32
+
+Checking harness check_usize...
+
+Status: SATISFIED\
+Description: "This may be true"\
+main.rs:32:1 in function check_usize
+
+Status: SATISFIED\
+Description: "This may also be true"\
+main.rs:32:1 in function check_usize
+
+Status: SATISFIED\
+Description: "v1 may be MAX"\
+main.rs:32:1 in function check_usize
+
+Status: SATISFIED\
+Description: "v1 may be MIN"\
+main.rs:32:1 in function check_usize
+
+Checking harness check_i128...
+
+Status: SATISFIED\
+Description: "This may be true"\
+main.rs:24:1 in function check_i128
+
+Status: SATISFIED\
+Description: "This may also be true"\
+main.rs:24:1 in function check_i128
+
+Status: SATISFIED\
+Description: "v1 may be MAX"\
+main.rs:24:1 in function check_i128
+
+Status: SATISFIED\
+Description: "v1 may be MIN"\
+main.rs:24:1 in function check_i128
+
+Checking harness check_i64...
+
+Status: SATISFIED\
+Description: "This may be true"\
+main.rs:23:1 in function check_i64
+
+Status: SATISFIED\
+Description: "This may also be true"\
+main.rs:23:1 in function check_i64
+
+Status: SATISFIED\
+Description: "v1 may be MAX"\
+main.rs:23:1 in function check_i64
+
+Status: SATISFIED\
+Description: "v1 may be MIN"\
+main.rs:23:1 in function check_i64
+
+Checking harness check_i8...
+
+Status: SATISFIED\
+Description: "This may be true"\
+main.rs:20:1 in function check_i8
+
+Status: SATISFIED\
+Description: "This may also be true"\
+main.rs:20:1 in function check_i8
+
+Status: SATISFIED\
+Description: "v1 may be MAX"\
+main.rs:20:1 in function check_i8
+
+Status: SATISFIED\
+Description: "v1 may be MIN"\
+main.rs:20:1 in function check_i8
+
+Checking harness check_isize...
+
+Status: SATISFIED\
+Description: "This may be true"\
+main.rs:25:1 in function check_isize
+
+Status: SATISFIED\
+Description: "This may also be true"\
+main.rs:25:1 in function check_isize
+
+Status: SATISFIED\
+Description: "v1 may be MAX"\
+main.rs:25:1 in function check_isize
+
+Status: SATISFIED\
+Description: "v1 may be MIN"\
+main.rs:25:1 in function check_isize
+
+Checking harness check_u16...
+
+Status: SATISFIED\
+Description: "This may be true"\
+main.rs:28:1 in function check_u16
+
+Status: SATISFIED\
+Description: "This may also be true"\
+main.rs:28:1 in function check_u16
+
+Status: SATISFIED\
+Description: "v1 may be MAX"\
+main.rs:28:1 in function check_u16
+
+Status: SATISFIED\
+Description: "v1 may be MIN"\
+main.rs:28:1 in function check_u16
+
+Checking harness check_i32...
+
+Status: SATISFIED\
+Description: "This may be true"\
+main.rs:22:1 in function check_i32
+
+Status: SATISFIED\
+Description: "This may also be true"\
+main.rs:22:1 in function check_i32
+
+Status: SATISFIED\
+Description: "v1 may be MAX"\
+main.rs:22:1 in function check_i32
+
+Status: SATISFIED\
+Description: "v1 may be MIN"\
+main.rs:22:1 in function check_i32
+
+Checking harness check_i16...
+
+Status: SATISFIED\
+Description: "This may be true"\
+main.rs:21:1 in function check_i16
+
+Status: SATISFIED\
+Description: "This may also be true"\
+main.rs:21:1 in function check_i16
+
+Status: SATISFIED\
+Description: "v1 may be MAX"\
+main.rs:21:1 in function check_i16
+
+Status: SATISFIED\
+Description: "v1 may be MIN"\
+main.rs:21:1 in function check_i16
+
+Checking harness check_u64...
+
+Status: SATISFIED\
+Description: "This may be true"\
+main.rs:30:1 in function check_u64
+
+Status: SATISFIED\
+Description: "This may also be true"\
+main.rs:30:1 in function check_u64
+
+Status: SATISFIED\
+Description: "v1 may be MAX"\
+main.rs:30:1 in function check_u64
+
+Status: SATISFIED\
+Description: "v1 may be MIN"\
+main.rs:30:1 in function check_u64
+
+Checking harness check_u8...
+
+Status: SATISFIED\
+Description: "This may be true"\
+main.rs:27:1 in function check_u8
+
+Status: SATISFIED\
+Description: "This may also be true"\
+main.rs:27:1 in function check_u8
+
+Status: SATISFIED\
+Description: "v1 may be MAX"\
+main.rs:27:1 in function check_u8
+
+Status: SATISFIED\
+Description: "v1 may be MIN"\
+main.rs:27:1 in function check_u8
+
+ ** 4 of 4 cover properties satisfied
+
+
+VERIFICATION:- SUCCESSFUL
+
+Complete - 12 successfully verified harnesses, 0 failures, 12 total.
