--- tests/expected/expect-fail/expected
@@ -1,2 +1,0 @@
-FAILURE\
-Description: "EXPECTED FAIL: x is not 5!"
