--- tests/kani/DynTrait/dyn_fn_once_fail.rs
@@ -1,19 +1,0 @@
-// Copyright Kani Contributors
-// SPDX-License-Identifier: Apache-2.0 OR MIT
-
-// Check that we can pass a dyn FnOnce pointer to a stand alone
-// function definition. Expected to fail because we are comparing
-// to an incorrect value.
-
-fn takes_dyn_fun(fun: Box<dyn FnOnce() -> u32>) -> u32 {
-    fun()
-}
-
-pub fn unit_to_u32() -> u32 {
-    5
-}
-
-#[kani::proof]
-fn main() {
-    kani::expect_fail(takes_dyn_fun(Box::new(&unit_to_u32)) == 3, "Wrong u32")
-}
