--- tests/ui/Property-Class-UI/expect_fail/expected
@@ -1,3 +1,0 @@
-main.expect_fail.1\
-Status: FAILURE
-Description: "EXPECTED FAIL: Blocked by assumption above."
