--- tests/ui/expect-fail-property-class/expected
@@ -1,3 +1,0 @@
-Check 1: main.expect_fail.1
-Status: FAILURE
-Description: "EXPECTED FAIL: Blocked by assumption above."
